refactor(client/v2): refactor of flags (#17306)
