+++ client/v2/autocli/flag/interface.go
@@ -1,9 +1,21 @@
 package flag
 
 import (
+	"context"
+
 	"github.com/spf13/pflag"
 	"google.golang.org/protobuf/reflect/protoreflect"
 )
+
+// Type specifies a custom flag type.
+type Type interface {
+	// NewValue returns a new pflag.Value which must also implement either
+	// SimpleValue or ListValue.
+	NewValue(context.Context, *Builder) Value
+
+	// DefaultValue is the default value for this type.
+	DefaultValue() string
+}
 
 // Value represents a single pflag.Value value.
 type Value interface {
