+++ client/v2/autocli/flag/maps/generic.go
@@ -1,4 +1,4 @@
-package flag
+package maps
 
 import (
 	"strings"
