+++ client/v2/autocli/flag/maps/map_int64_to_value.go
@@ -1,4 +1,4 @@
-package flag
+package maps
 
 import (
 	"strconv"
