+++ client/v2/autocli/testdata/help-deprecated-msg.golden
@@ -8,9 +8,9 @@ Flags:
       --a-bool                                                               
       --a-coin cosmos.base.v1beta1.Coin                                      
       --a-message testpb.AMessage (json)                                     
-      --a-validator-address bech32 account address key name                  
+      --a-validator-address bech32 account address                           
   -a, --account-number uint                                                  The account number of the signing account (offline mode only)
-      --an-address bech32 account address key name                           
+      --an-address bech32 account address                                    
       --an-enum Enum (unspecified | one | two | five | neg-three)             (default unspecified)
       --aux                                                                  Generate aux signer data instead of sending a tx
       --bools bools                                                           (default [])
