+++ client/v2/autocli/testdata/help-deprecated.golden
@@ -5,10 +5,10 @@ Usage:
 Flags:
       --a-bool                                                               
       --a-coin cosmos.base.v1beta1.Coin                                      
-      --a-consensus-address bech32 account address key name                  
+      --a-consensus-address bech32 account address                           
       --a-message testpb.AMessage (json)                                     
-      --a-validator-address bech32 account address key name                  
-      --an-address bech32 account address key name                           
+      --a-validator-address bech32 account address                           
+      --an-address bech32 account address                                    
       --an-enum Enum (unspecified | one | two | five | neg-three)             (default unspecified)
       --bools bools                                                           (default [])
       --bz binary
