+++ client/v2/autocli/testdata/help-echo.golden
@@ -12,10 +12,10 @@ echo 1 abc {}
 Flags:
       --a-bool                                                               
       --a-coin cosmos.base.v1beta1.Coin                                      some random coin
-      --a-consensus-address bech32 account address key name                  
+      --a-consensus-address bech32 account address                           
       --a-message testpb.AMessage (json)                                     
-      --a-validator-address bech32 account address key name                  
-      --an-address bech32 account address key name                           
+      --a-validator-address bech32 account address                           
+      --an-address bech32 account address                                    
       --an-enum Enum (unspecified | one | two | five | neg-three)             (default unspecified)
       --bools bools                                                           (default [])
       --bz binary                                                            some bytes
